+++ lib/src/draw/draw_circle.dart
@@ -1,6 +1,9 @@
+import 'dart:math';
+
 import '../color/channel.dart';
 import '../color/color.dart';
 import '../image/image.dart';
+import '../util/math_util.dart';
 import '_calculate_circumference.dart';
 import 'draw_pixel.dart';
 
@@ -11,8 +14,39 @@ Image drawCircle(Image image,
     required int y,
     required int radius,
     required Color color,
+    bool antialias = false,
     Image? mask,
     Channel maskChannel = Channel.luminance}) {
+  if (antialias) {
+    void drawPixel4(int x, int y, int dx, int dy, num alpha) {
+      drawPixel(image, x + dx, y + dy, color,
+          alpha: alpha, mask: mask, maskChannel: maskChannel);
+
+      drawPixel(image, x - dx, y + dy, color,
+          alpha: alpha, mask: mask, maskChannel: maskChannel);
+
+      drawPixel(image, x + dx, y - dy, color,
+          alpha: alpha, mask: mask, maskChannel: maskChannel);
+
+      drawPixel(image, x - dx, y - dy, color,
+          alpha: alpha, mask: mask, maskChannel: maskChannel);
+    }
+
+    final radiusSqr = radius * radius;
+    final quarter = (radius / sqrt2).round();
+    for (var i = 0; i < quarter; ++i) {
+      final j = sqrt(radiusSqr - (i * i));
+      final frc = fract(j);
+      final flr = j.floor();
+      drawPixel4(x, y, i, flr, 1 - frc);
+      drawPixel4(x, y, i, flr + 1, frc);
+      drawPixel4(x, y, flr, i, 1 - frc);
+      drawPixel4(x, y, flr + 1, i, frc);
+    }
+
+    return image;
+  }
+
   final points = calculateCircumference(image, x, y, radius);
   for (final pt in points) {
     drawPixel(image, pt.xi, pt.yi, color, mask: mask, maskChannel: maskChannel);